requirements.txt
requests
openpyxl
Pillow
tinify
supabase
google-api-python-client
google-auth
python-dotenv
